album+ track

+++ app/Http/Controllers/AlbumController.php
@@ -5,23 +5,18 @@ namespace App\Http\Controllers;
 use App\Models\Album;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Storage;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 
 class AlbumController extends Controller
 {
+    use AuthorizesRequests;
+    
     /**
      * Display a listing of the resource.
      */
     public function index()
     {
-        // return Album::with('user')->get();
-
-        $user_id = auth()->user()->id;
-        $albums = Album::where('user_id', $user_id)->get();
-
-        return response()->json([
-            'albums' => $albums,
-            'status' => true
-        ]);
+        return Album::with(['user', 'genre'])->latest()->get();
     }
 
     /**
@@ -29,24 +24,23 @@ class AlbumController extends Controller
      */
     public function store(Request $request)
     {
-       $data = $request->validate([
+      $validated = $request->validate([
             'title' => 'required|string',
-            // 'user_id' => 'required|exists:users,id'
+            'description' => 'nullable|string',
+            'cover_image' => 'nullable|image',
+            'genre_id' => 'required|exists:genres,id',
+            // 'release_date'=> 'nullable|date',
         ]);
 
-        $data['user_id'] = auth()->user()->id;
-
-        if($request->hasFile('cover_image')) {
-            $data['cover_image'] = $request->file('cover_image')->store('album_covers', 'public');
-            // $data['cover_image'] = '/storage/' . $path;
+        if ($request->hasFile('cover_image')) {
+            $validated['cover_image'] = $request->file('cover_image')->store('albums', 'public');
         }
 
-        Album::create($data);
+        $validated['user_id'] = auth()->id();
 
-        return response()->json([
-            'status' => true,
-            'message' => 'Album created successfully'
-        ]);
+        $album = Album::create($validated);
+
+        return response()->json($album, 201);
     }
 
     /**
@@ -54,11 +48,7 @@ class AlbumController extends Controller
      */
     public function show(Album $album)
     {
-        return response()->json([
-            'album' => $album,
-            'status' => true,
-            'message' => 'Album Data Found'
-        ]);
+       return $album->load(['user', 'tracks', 'genre']);
     }
 
     /**
@@ -66,23 +56,17 @@ class AlbumController extends Controller
      */
     public function update(Request $request, Album $album)
     {
-        $data = $request->validate([
-            'title' => 'required',
-        ]);
+       $this->authorize('update', $album);
 
-        if($request->hasFile('cover_image')) {
-           if($album->cover_image) {
-                Storage::disk("public")->delete($album->cover_image);
-            }
-            $data['cover_image'] = $request->file('cover_image')->store('album_covers', 'public');
+        $data = $request->only(['title', 'description', 'genre_id']);
+
+        if ($request->hasFile('cover_image')) {
+            $data['cover_image'] = $request->file('cover_image')->store('albums', 'public');
         }
 
         $album->update($data);
 
-        return response()->json([
-            'status' => true,
-            'message' => 'Album updated successfully'
-        ]);
+        return response()->json($album);
 
     }
 
@@ -91,11 +75,10 @@ class AlbumController extends Controller
      */
     public function destroy(Album $album)
     {
+        $this->authorize('delete', $album);
+
         $album->delete();
 
-        return response()->json([
-            'status' => true,
-            'message' => 'Album deleted successfully'
-        ]);
+        return response()->json(null, 204);
     }
 }
